[cost_map] large refactor and cleanup
