+++ cost_map_core/include/cost_map_core/iterators/iterators.hpp
@@ -1,17 +1,13 @@
-/*
- * iterators.hpp
- *
- *  Created on: Sep 22, 2014
- *      Author: Péter Fankhauser
- *   Institute: ETH Zurich, Autonomous Systems Lab
+/**
+ * @file /cost_map_core/include/cost_map_core/iterators/iterators.hpp
  */
 
 #pragma once
 
-#include "CircleIterator.hpp"
-#include "CostMapIterator.hpp"
-#include "EllipseIterator.hpp"
-#include "LineIterator.hpp"
-#include "PolygonIterator.hpp"
-#include "SpiralIterator.hpp"
-#include "SubmapIterator.hpp"
+#include "circle_iterator.hpp"
+#include "costmap_iterator.hpp"
+#include "ellipse_iterator.hpp"
+#include "line_iterator.hpp"
+#include "polygon_iterator.hpp"
+#include "spiral_iterator.hpp"
+#include "submap_iterator.hpp"
