+++ cost_map_core/include/cost_map_core/operators/inflation.hpp
@@ -12,7 +12,7 @@
 ** Includes
 *****************************************************************************/
 
-#include "../CostMap.hpp"
+#include "../cost_map.hpp"
 #include <limits>
 #include <queue>
