+++ cost_map_core/src/lib/submap_geometry.cpp
@@ -1,13 +1,10 @@
-/*
- * SubmapGeometry.cpp
- *
- *  Created on: Aug 18, 2015
- *      Author: Péter Fankhauser
- *   Institute: ETH Zurich, Autonomous Systems Lab
+/**
+ * @file src/lib/submap_geometry.cpp
  */
 
 #include <grid_map_core/GridMapMath.hpp>
-#include "../../include/cost_map_core/SubmapGeometry.hpp"
+#include "../../include/cost_map_core/submap_geometry.hpp"
+#include "../../include/cost_map_core/cost_map.hpp"
 
 
 namespace cost_map {
