+++ cost_map_core/test/CMakeLists.txt
@@ -1,17 +1,18 @@
+###############################################################################
+# GTest setup
+###############################################################################
+
+macro(cost_map_core_add_gtest TESTNAME)
+  catkin_add_gtest(cost_map_core_test_${TESTNAME} ${TESTNAME}.cpp)
+  target_link_libraries(cost_map_core_test_${TESTNAME} ${PROJECT_NAME} ${catkin_LIBRARIES})
+endmacro(cost_map_core_add_gtest)
+
 ##############################################################################
 # Tests
 ##############################################################################
 
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -pthread")
-# Add gtest based cpp test target and link libraries
-catkin_add_gtest(${PROJECT_NAME}-test
-  test_cost_map_core.cpp
-  CostMapTest.cpp
-  EllipseIteratorTest.cpp
-  SubmapIteratorTest.cpp
-  PolygonIteratorTest.cpp
-)
-if(TARGET ${PROJECT_NAME}-test)
-  target_link_libraries(${PROJECT_NAME}-test ${PROJECT_NAME})
-endif()
-
+cost_map_core_add_gtest(cost_map_core)
+cost_map_core_add_gtest(cost_map)
+cost_map_core_add_gtest(ellipse_iterator)
+cost_map_core_add_gtest(submap_iterator)
+cost_map_core_add_gtest(polygon_iterator)
