+++ cost_map_visualisations/src/lib/occupancy_grid.cpp
@@ -5,7 +5,7 @@
 ** Includes
 *****************************************************************************/
 
-#include <cost_map_core/CostMap.hpp>
+#include <cost_map_core/cost_map.hpp>
 #include <cost_map_ros/converter.hpp>
 #include <nav_msgs/OccupancyGrid.h>
 #include "../../include/cost_map_visualisations/occupancy_grid.hpp"
